Refactor v1 router to use separated docs route
Documentation should be its own line of concern. Moved all related
resources and endpoints to be served under its own /docs subpath.
Also did a quick documentation cleanup pass and ensured that all html
links were wrapped in a @link jsdoc annotation.
