+++ app/src/docs/docs.js
@@ -14,7 +14,7 @@ const docs = {
       </style>
     </head>
     <body>
-      <redoc spec-url='/api/${version}/api-spec.yaml' />
+      <redoc spec-url='/api/${version}/docs/api-spec.yaml' />
       <script src="https://cdn.jsdelivr.net/npm/redoc@next/bundles/redoc.standalone.js"></script>
     </body>
   </html>`
