+++ app/src/routes/v1/index.js
@@ -1,21 +1,7 @@
-const config = require('config');
-const fs = require('fs');
-const path = require('path');
 const router = require('express').Router();
-const yaml = require('js-yaml');
-
 const { currentUser } = require('../../middleware/authentication');
 
 router.use(currentUser);
-
-/** Gets the OpenAPI specification */
-const getSpec = () => {
-  const rawSpec = fs.readFileSync(path.join(__dirname, '../../docs/v1.api-spec.yaml'), 'utf8');
-  const spec = yaml.load(rawSpec);
-  spec.servers[0].url = '/api/v1';
-  spec.components.securitySchemes.OpenID.openIdConnectUrl = `${config.get('keycloak.serverUrl')}/realms/${config.get('keycloak.realm')}/.well-known/openid-configuration`;
-  return spec;
-};
 
 // Base v1 Responder
 router.get('/', (_req, res) => {
@@ -29,21 +15,8 @@ router.get('/', (_req, res) => {
   });
 });
 
-/** OpenAPI Docs */
-router.get('/docs', (_req, res) => {
-  const docs = require('../../docs/docs');
-  res.send(docs.getDocHTML('v1'));
-});
-
-/** OpenAPI YAML Spec */
-router.get('/api-spec.yaml', (_req, res) => {
-  res.status(200).type('application/yaml').send(yaml.dump(getSpec()));
-});
-
-/** OpenAPI JSON Spec */
-router.get('/api-spec.json', (_req, res) => {
-  res.status(200).json(getSpec());
-});
+/** Documentation Router */
+router.use('/docs', require('./docs'));
 
 /** Object Router */
 router.use('/object', require('./object'));
