+++ app/tests/unit/routes/v1/object.spec.js
@@ -9,7 +9,7 @@ const { objectController } = require('../../../../src/controllers');
 const basePath = '/api/v1/object';
 const app = expressHelper(basePath, router);
 
-// Mock config library - @see https://stackoverflow.com/a/64819698
+// Mock config library - @see {@link https://stackoverflow.com/a/64819698}
 jest.mock('config');
 
 describe(`POST ${basePath}`, () => {
