+++ app/tests/unit/services/storage.spec.js
@@ -24,7 +24,7 @@ const s3ClientMock = mockClient(service._s3Client);
 jest.mock('@aws-sdk/s3-request-presigner', () => ({
   getSignedUrl: jest.fn()
 }));
-// Mock config library - @see https://stackoverflow.com/a/64819698
+// Mock config library - @see {@link https://stackoverflow.com/a/64819698}
 jest.mock('config');
 
 beforeEach(() => {
